+++ drizzle/0002_slim_dragon_man.sql
@@ -1,0 +1,4 @@
+ALTER TYPE "setting_keys" ADD VALUE 'global_channel_id';--> statement-breakpoint
+ALTER TYPE "setting_keys" ADD VALUE 'join_role_id';--> statement-breakpoint
+ALTER TYPE "setting_keys" ADD VALUE 'join_to_talk_channel_id';--> statement-breakpoint
+ALTER TYPE "setting_keys" ADD VALUE 'temp_voice_channels_category_id';