+++ src/classes/Button.ts
@@ -1,3 +1,5 @@
+import { ButtonInteraction } from 'discord.js';
+
 import IButton from '../interfaces/IButton';
 import Client from './Client';
 
@@ -11,7 +13,7 @@ export default class Button implements IButton {
     this.id = id;
   }
 
-  Execute(): void {
-    throw new Error('Method not implemented.');
+  Execute(interaction: ButtonInteraction): void {
+    throw new Error(`Method not implemented in ${interaction.id} button.`);
   }
 }